rename fiber source and example files
