+++ src/lib.rs
@@ -243,13 +243,13 @@ mod unwind;
 
 mod arch;
 mod coroutine;
-mod execution;
+mod fiber;
 pub mod stack;
 pub mod trap;
 mod util;
 
 pub use coroutine::*;
-pub use execution::*;
+pub use fiber::*;
 
 #[cfg(test)]
 mod tests;
